start adding voice recorder widget
